Remove rack_dyn_draw and just use DynamicDisplay
