+++ firmware/src/gui/dyn_display.hh
@@ -1,27 +1,27 @@
 #pragma once
-#include "gui/dyn_element/base_dyn_draw.hh"
-#include "gui/dyn_element/native_dyn_draw.hh"
+#include "gui/dyn_display_drawer.hh"
 #include "patch_play/patch_playloader.hh"
 #include <memory>
 
 namespace MetaModule
 {
 
-class DynamicElementDraw {
+class DynamicDisplay {
 public:
-	DynamicElementDraw(PatchPlayLoader &patch_playloader)
+	DynamicDisplay(PatchPlayLoader &patch_playloader)
 		: patch_playloader{patch_playloader} {
 	}
 
-	void prepare_module(std::string_view slug, unsigned module_id, lv_obj_t *module_canvas, unsigned px_per_3U) {
+	bool prepare_module(std::string_view slug, unsigned module_id, lv_obj_t *module_canvas, unsigned px_per_3U) {
 		pr_trace("DynamicElementDraw: Prepare canvas for %s, id %u, pxp3u %u\n", slug.data(), module_id, px_per_3U);
 
 		if (auto module = patch_playloader.get_plugin_module<CoreProcessor>(module_id)) {
-			drawer = std::make_unique<DynDraw>(module, slug);
+			drawer = std::make_unique<DynamicDisplayDrawer>(module, slug);
 			drawer->prepare(module_canvas, px_per_3U);
-		} else {
-			pr_dbg("Module idx %d is no a CoreProcessor?\n", module_id);
+			return true;
 		}
+
+		return false;
 	}
 
 	void draw() {
@@ -32,17 +32,14 @@ public:
 
 	void blur() {
 		if (drawer) {
-			drawer->blur();
+			pr_dbg("DynamicElementDraw::blur()\n");
+			drawer.reset();
 		}
-	}
-
-	bool is_active() {
-		return drawer ? true : false;
 	}
 
 private:
 	PatchPlayLoader &patch_playloader;
-	std::unique_ptr<DynDraw> drawer{};
+	std::unique_ptr<DynamicDisplayDrawer> drawer{};
 };
 
 } // namespace MetaModule
